refactor(diff-viewer): entrypoint を module に移動
